backend/requirements.txt
# Core Web Framework
fastapi==0.121.2
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# PDF Processing
PyMuPDF==1.26.6

# Image Processing
Pillow==12.0.0

# Machine Learning / YOLO
torch==2.9.1
ultralytics==8.3.228

# Required by ultralytics and other packages
numpy==2.2.6

# Optional but recommended for better performance
# opencv-python==4.9.0.80  # Uncomment if you need additional CV operations
